refactor(logger): sanitize ロジックを lib/utils/sanitize-error.ts に分離
CI で zenn-api.test.ts が `(0 , _logger.sanitizeErrorMessage) is not a function`
で失敗していた問題を修正。

原因: 53 箇所の test ファイルが `jest.mock('@/lib/logger')` で logger を
mock しているが、いずれも default export (logger オブジェクト) のみを
mock しており、新規追加した named export `sanitizeErrorMessage` は
undefined になる。zenn-api.test.ts は logEnrichmentError → classifyEnrichmentError
→ sanitizeErrorMessage の呼び出し経路を踏むため failure。

修正:
- lib/utils/sanitize-error.ts を新設し sanitizeError / sanitizeErrorMessage を
  pure 関数として独立モジュール化 (logger 非依存)
- lib/logger.ts は新モジュールから import して再 export することで
  既存の `@/lib/logger` 経由の import パスを後方互換維持
- error-classifier.ts は logger を経由せず直接 sanitize-error から import

これにより `@/lib/logger` をモックしているテストでもサニタイズ関数は
実体が解決され、本問題が再発しない。

Refs #605
